requirements.txt
openai>=1.66.5
openai-agents>=0.0.11
yfinance>=0.2.0
pydantic>=2.0.0
python-dotenv>=0.21.0
streamlit>=1.35.0
plotly>=5.20.0